+++ example/lib/main.dart
@@ -49,7 +49,7 @@ class _MyHomePageState extends State<MyHomePage> {
           ListTile(
             title: Text(
               'AzureAD OAuth',
-              style: Theme.of(context).textTheme.headline5,
+              style: Theme.of(context).textTheme.headlineSmall,
             ),
           ),
           ListTile(
@@ -68,7 +68,12 @@ class _MyHomePageState extends State<MyHomePage> {
               },
             ),
           ListTile(
-            leading: Icon(Icons.delete),
+            leading: Icon(Icons.data_array),
+            title: Text('HasCachedAccountInformation'),
+            onTap: () => hasCachedAccountInformation(),
+          ),
+          ListTile(
+            leading: Icon(Icons.logout),
             title: Text('Logout'),
             onTap: () {
               logout();
@@ -103,9 +108,21 @@ class _MyHomePageState extends State<MyHomePage> {
     );
     var accessToken = await oauth.getAccessToken();
     if (accessToken != null) {
+      ScaffoldMessenger.of(context).hideCurrentSnackBar();
       ScaffoldMessenger.of(context)
           .showSnackBar(SnackBar(content: Text(accessToken)));
     }
+  }
+
+  void hasCachedAccountInformation() async {
+    var hasCachedAccountInformation = await oauth.hasCachedAccountInformation;
+    ScaffoldMessenger.of(context).hideCurrentSnackBar();
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content:
+            Text('HasCachedAccountInformation: $hasCachedAccountInformation'),
+      ),
+    );
   }
 
   void logout() async {