+++ lib/model/config.dart
@@ -1,3 +1,4 @@
+import 'package:aad_oauth/model/cache_location.dart';
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
@@ -111,6 +112,13 @@ class Config {
   /// Azure Active Directory B2C provides business-to-customer identity as a service.
   final bool isB2C;
 
+  /// When using Azure AD B2C with a custom domain or Azure Front Door, 
+  /// the custom domain URL must be used instead of the default login.microsoftonline.com URL.
+  /// This will change the issuer of the token to the custom domain URL.
+  /// Example: https://account.examplecompany.com/01234567-89ab-cdef-0123-456789abcdef.
+  /// More information can be found here: https://learn.microsoft.com/en-us/azure/active-directory-b2c/custom-domain.
+  final String? customDomainUrlWithTenantId;
+
   /// Flag whether to use a stub implementation for unit testing or not
   bool isStub;
 
@@ -126,8 +134,29 @@ class Config {
   /// android storage options for shared preferences - defaults to encrypting shared prefs
   AndroidOptions aOptions;
 
+  /// Cache location used when authenticating with a web client.
+  /// "CacheLocation.localStorage" - Local browser storage (default)
+  /// "CacheLocation.sessionStorage" - Session context
+  /// "CacheLocation.memoryStorage" - Memory only
+  CacheLocation cacheLocation;
+
   /// Loader Widget (before load web page)
   Widget loader;
+
+  /// Origin header parameter for [TokenRequestDetails] and [TokenRefreshRequestDetails]
+  String? origin;
+
+  /// Support for custom url parameters for dynamic UI support
+  /// View docs on Azure B2C:
+  /// https://learn.microsoft.com/en-us/azure/active-directory-b2c/claim-resolver-overview#dynamic-ui-customization
+  final Map<String, String> customParameters;
+
+  /// Sign-out with a redirect
+  /// On Azure logout process, it'll redirect the user to this url
+  /// By this we can verify the user is logged out successfully
+  /// View docs:
+  /// https://learn.microsoft.com/en-us/azure/active-directory/develop/scenario-spa-sign-in?tabs=javascript2#tabpanel_4_javascript2
+  String? postLogoutRedirectUri;
 
   /// Determine an appropriate redirect URI for AAD authentication.
   /// On web, it is the location that the application is being served from.
@@ -167,6 +196,7 @@ class Config {
     this.clientSecret,
     this.resource,
     this.isB2C = false,
+    this.customDomainUrlWithTenantId,
     this.loginHint,
     this.domainHint,
     this.codeVerifier,
@@ -174,15 +204,25 @@ class Config {
     this.isStub = false,
     this.loader = const SizedBox(),
     AndroidOptions? aOptions,
+    CacheLocation? cacheLocation,
     this.navigatorKey,
+    this.origin,
+    this.customParameters = const {},
+    String? postLogoutRedirectUri,
     this.appRouter,
   })  : assert(navigatorKey != null || appRouter != null),
         authorizationUrl = isB2C
-            ? 'https://$tenant.b2clogin.com/$tenant.onmicrosoft.com/$policy/oauth2/v2.0/authorize'
+            ? (customDomainUrlWithTenantId == null 
+              ? 'https://$tenant.b2clogin.com/$tenant.onmicrosoft.com/$policy/oauth2/v2.0/authorize'
+              : '$customDomainUrlWithTenantId/$policy/oauth2/v2.0/authorize')
             : 'https://login.microsoftonline.com/$tenant/oauth2/v2.0/authorize',
         tokenUrl = isB2C
-            ? 'https://$tenant.b2clogin.com/$tenant.onmicrosoft.com/$policy/oauth2/v2.0/token'
+            ? (customDomainUrlWithTenantId == null
+              ? 'https://$tenant.b2clogin.com/$tenant.onmicrosoft.com/$policy/oauth2/v2.0/token'
+              : '$customDomainUrlWithTenantId/$policy/oauth2/v2.0/token')
             : 'https://login.microsoftonline.com/$tenant/oauth2/v2.0/token',
+        postLogoutRedirectUri = postLogoutRedirectUri,
         aOptions = aOptions ?? AndroidOptions(encryptedSharedPreferences: true),
+        cacheLocation = cacheLocation ?? CacheLocation.localStorage,
         redirectUri = redirectUri ?? getDefaultRedirectUri();
 }